product-admin/requirements.txt
Django>=4.0,<5.0
psycopg2>=2.9.3
django-grappelli>=3.0.3,<4.0
kafka-python>=2.0.2,<3.0